backend/requirement.txt
django
djangorestframework
psycopg2-binary
openai